requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.5.0
astroid==3.3.3
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mock==5.1.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pylint==3.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
requests-mock==1.12.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
